Updated OIDC client with fields necessary for UAA
| Name                   | Attribute                      | Description                                                       |
|------------------------|--------------------------------|-------------------------------------------------------------------|
| Client Secret          | `vmwOidcClientSecret`          | A password field                                                  |
| Authorities            | `vmwOidcClientAuthorities`     | A list of authorities granted to the client                       |
| Resource IDs           | `vmwOidcResourceIds`           | A list of resources that the client can access                    |
| Scopes                 | `vmwOidcScopes`                | The scopes that are assigned to a client                          |
| Auto Approved Scopes   | `vmwOidcAutoApproveScopes`     | The scopes for which the client does not require user approval    |
| Authorized Grant Types | `vmwOidcAuthorizedGrantTypes`  | The grant types for which this client is authorized               |
| Additional Information | `vmwOidcAdditionalInformation` | Additional information not required for vanilla OAuth (JSON blob) |

OIDC clients no longer require a redirect URI in order to support
redirectless clients (e.g. the default 'cf' and 'admin' uaa clients).
The vmdir schema is unfortunately stuck with the requirement as we
cannot convert the 'MUST' field into a 'MAY' field (at least at the
moment), so we substitute in a sentinel value of "N/A". This may be
removed once vmdir supports the transition.

Client secrets are encoded using AES with a UTF-16 character set. The
encoding mechanism isn't particularly optimal as it involves a lot of
recreating of clients, but there didn't seem to be a very clean
mechanism for passing along the Crypto/tenant key to the location where
encrypt/decrypt needed to occur.

Bumped UAA spec to 013666bdd (latest).

Change-Id: Ic079bbf9b536d0e61b042129eda8aff20e839450