Fixed issues with HOK token signatures related to request redirect and null entities.
When HOK tokens were used with a port number, the client included the
port number as part of the signed request while the server does not.
This change sanitizes the URI by removing the scheme, userInfo, host,
and port, leaving only the path, query and fragment as part of the
signed material.

When provided with a null entity the client was signing an entity
string of "null" while the server would correctly leave the string
empty ("").

Bugs: #1837945, #1837963

Change-Id: I308ba199148e16ec0c12e6740a8dcdbe01f218d4