Support vmkdcd listeing on both IPv4 and IPv6 interfaces
Changed vmkdcd to bind on both IPv4 and IPv6 wildcard addresses, both TCP
and UDP. bind() will fail should IPv4 or IPv6 not be available, in
which case that socket fd is set to -1, and is not used in the
select() FD_SET mask.

Test iAcceptSock for -1 before use in listen() or adding to FD_SET.

Fixed if() test around listen() found by JB. Found/fixed similar
bad if() checks around FD_SET() calls for the 4 types of socket descriptors.

Address JB review comment

This minor refactoring avoids the possibility of a TCP connection
being accepted but lost, as the accepted socket is over written.
That would result in a socket leak, and hung client connections.
In the case of UDP, pending "connections" would have just been dropped,
and retried.

Change-Id: I7a25510779ae73dd9885cebcee110f51987a2238