Merge UAA into the uber-WAR
ROOT.war contains copies of the configuration files, resources, and
Thymeleaf template files owned by UAA and the CloudFoundry Identity
Server. Because of the construction of the uber-WAR many hyperlinks and
redirects which rely on the servlet context do not work - because we
deploy as ROOT the servlet context is considered to be the empty string
rather than the servlet mapping that we configure in web.xml. To
accomodate this we override the ThymeleafViewResolver and RedirectView
with Servlet-Mapping-Aware variations. Thymeleaf-based hyperlinks still
require that we prepend the servlet mapping because the servletContext
is not populated.

In the future we may be able to create an additional filter that wraps
the request (HttpServletRequestWrapper) and supplies a different context
path based on the servlet mapping. Unfortunately the attempts to do so
created issues with redirect loops that weren't worth solving
immediately.

Changes:

* Refactored OIDC and WebSSO into separate servlets so that the root
  appServlet no longer interfered with requests.
* Took copies of the UAA config files in order to customize them for the
  uber-WAR.
* Created RedirectViewWithServletMapping and
  ThymeleafViewResolverWithServletMapping to supply servlet-mapping
  aware redirection.
* Updated template links to use the servlet mapping.
* Updated HSQLDB validationQuery to a supported one to stop the
  intermittent JDBC validation failures.
* Added default target URL for
  UaaSavedRequestAwareAuthenticationSuccessHandler to properly redirect
  in some situations (after logging out and logging back in most
  noticeably).
* ROOT now contains additional Java classes that are meant specifically
  for its own operation:
    * Moved STSLogDiagnosticsFilter out of the Diagnostics lib and into
      ROOT's build path since it only exists for mapping servlet paths
      to log files.

Change-Id: Icccecafc60401527ca536f49b249c77d03930572