Support different CLDAP client queries
The original CLDAP "ping" implementation assumed there was only one
query type: netlogon. This is wrong. Arbitrary LDAP filters can
be applied to CLDAP requests. This change is a compromise. matchBlob()
searches the ASN.1-encoded data for specific strings in the
match term. This is a baby step towards actually ber_scanf()/ber_read()
and properly treating these queries as actual LDAP requests.
This was an expedient way to distinguish between different query
types, and reply correctly. This is intended to be replaced later
with a better implementation.

Change-Id: If237747d245e394d7e3d02e227cb0398516f76c6