Added SCIM filtering for users, groups and clients
Enabled the PassthroughPasswordEncoder so that clients created
through REST IDM still work with UAA (and UAAC).

Fixed NPE when fetching GC connection in ActiveDirectoryProvider.

Fixed potential memory leak in ActiveDirectoryProvider,
LdapWithAdMappingsProvider and LdapProvider centered around the
closing of multiple messages retrieved from an LDAP query.

In order to perform SCIM filtering, we convert from the generic
SCIMFilter object into an LDAP search query, much the same way as is
done in UAA (the general structure of search query converter, attribute
mapping, etc. all stem from the UAA implementation).

Example:

  "username pr"                 => "(cn=*)"
  "username eq \"james\""       => "(cn=james)"
  "(username eq \"james\"
     and
   name.familyname eq \"bond\") => "(&(cn=james)(sn=bond))"

SCIM filtering is not performed manually but leverages the identity
source's query mechanism - this means that certain providers simply do
not support certain query types. For example, VMwareDirectoryProvider
is unwilling to filter on the 'groups' attribute of a user. To get
around this, each provider can choose to implement a non-action mapping
that simply leaves out that clause. If the attribute does not have a
defined mapping it will instead choose to throw an
AttributeUndefinedException that will halt query construction.

Example:

  converter.setMapping("id", "upn");
  converter.setMapping("groups", (sb, type, value) -> {});

  "groups co \"Administrator\" => ""
  "id eq \"james@mi6.com\"
    and
   groups co \"Administrator\"" => "(&(upn=james@mi6.com))"
  "username eq \"james\"" => AttributeUndefinedException

Generally supported attributes (consult the provider for full details):

  * User Schema: https://tools.ietf.org/html/rfc7643#section-4.1
    * id                => UserAttributePrincipalName &
                           UserAttributeAccountName
    * externalId        => UserAttributeCommonName
    * username          => UserAttributeAccountName
    * name.familyname   => UserAttributeLastName
    * name.givenname    => UserAttributeFirstName
    * displayname       => UserAttributeDisplayName
    * emails            => UserAttributeEmail
    * groups            => UserAttributeMemberOf

  * Group Schema: https://tools.ietf.org/html/rfc7643#section-4.2
    * id                => GroupAttributeAccountName
    * externalId        => GroupAttributeAccountName
    * displayname       => GroupAttributeAccountName
    * members           => GroupAttributeMembersList

Change-Id: I625338b376960d593cf220db959a6652e6f7125e