Add UAA-compliant OIDC Clients to REST
Added the UAA-specific fields to the OIDCClientMetadataDTO.

Modified the OIDC Client creation endpoint to accept both an
OIDCClientMetadataDTO and an OIDCClientDTO so that we can create clients
where we specify the clientId explicitly (rather than forcing it to be
generated).

Created a Ruby script that utilizes the REST endpoints to establish
fixtures of Users, Groups, and Clients.

Added a pcfFixtures.yml file that creates the initial users, groups, and
clients necessary for the PCF demo.

Change-Id: Idc36e17661c7af7efef459f4ed9f137bb98db1e2