Solve hung lightwave services due to unresponsive dcerpcd
Do not register a service's RPC interfaces with dcerpcd.
All lightwave services use well-known ports. dcerpcd
is a portmapper, i.e. return the IP:[port] for an
RPC interface specification. Because of firewall rules
it is not possible to use ephemerial RPC ports, so no
component queries dcerpcd for a service for an RPC
interface. Since one deployment environment has revealed a
latent bug in either dcerpcd or libdcerpcd.so, don't register
with dcerpcd.

Comment out calls to VmXXXRpcEpRegister as well from server.c,
as per brownj comment.

Change-Id: Ibdf77eb65123afe11c0c19b525f7a3da25374ce1