Implemented creation of reverse zone with reverse query support
Added user capability to create a reverse zone and add,query and
delete records from the same. Previous implementation did not allow
the user to create a reverse zone.

Default reverse zones in-addr.arpa and ip6.arpa are created at
initialization.

Tested the code for adding, querying and deleting PTR records for ip4
and ip6 in both the default reverse zones and user-created reverse
zones. The tests were successful on both RPC port as well as DNS port
(port 53) verified through cli commands and tools like nslookup, dig
respectively.

Unit testing was done for the code by adding, querying and deleting a
PTR record. The file is included in the test directory in vmdns as TestPtr.c

The scripts under vmdns/test/scripts have been included for testing and
debugging purposes. They include both positive as well as negative tests
for the user to gain an understanding of reverse zones and their lookups

Change-Id: Id0be8f06e6e5cdda6a2c34fc6b6f284e2a8ebdcf