CNAME resolution for A and AAAA queries
Implemented the CNAME resolution for A and AAAA queries. In the previous
implementation, queries like A and AAAA made for a name containing only
a CNAME record resulted in error.

In this implementation, the problem mentioned above is resolved and
queries are made using the target name of the CNAME record present for
the rr type of the query. Hence multiple levels of indirections through
CNAME can be done before the record matching the type of query can be
obtained. We have currently limited the number of indirections to 5;
this can be changed as per required.

Also added error checks and bails for the condition that CNAME record
should not co-exist with any other record type. The changes are
reflected in VmDnsIsUpdatePermitted function.

Changes made:
- Edited the VmDnsSrvQueryRecords function to handle CNAME records
- Moved the functionality to get a record from the cache or the store to
  another function named VmDnsSrvGetRecords
- Created VmDnsGetLinkedRecords function which handles the queries
  needed to be made for CNAME records
- Modified the VmDnsIsUpdatePermitted function for proper error checks.

Testing:
The testing was done to check if A and AAAA queries for records containing
only the CNAME record were successful. Both of them resulted in success
on RPC as well as DNS ports. Negative tests were also done to make sure
that CNAME record could co-exist with any other record. Testing was also
done to ensure that upto 5 levels of indirections are supported
currently, and the code fails with "ERROR_NOT_SUPPORTED" for higher
number of indirections, thereby keeping a cap on the maximum response
time for a query.

Change-Id: Ibc5e1e51822d8f72ea939e064b0269fb954d6705