Implemented Round Robin Algorithm for Load Balancing
In the previous implementation, the records were returned in the order
that they were stored. Hence, when a name entry with multiple records
were queried, the records were always returned in the same order.

In the current implementation, round robin algorithm is used to permute
between the records each time the corresponding entry is queried. This
is helpful for load balancing as users usually use the first entry in
the response in case of multiple response entries.

The index and the record type are stored in the name entry structure
when a record is queried. Both these parameters are initialized when the
name entry is created in cache. If subsequently the same type is
queeried, the parameters from the structure are used, and if a different
type is queried, the default type and index values are used, and the
parameters in the name entry are set according to the current query.
This solves the issues when multiple types are present in the same name
entry.

Changes made:
 -Modified the name entry structure to have two additional entries, one
  for round robin index and the other for round robin type
 -Modified the function which is used to get records from the name entry
  to support round robin
 -Added a function which permutes the record list according to round robin
  algorithm

The code was tested for A and AAAA records. The algorithm works queries
on both, the RPC port and the DNS port. The case for multiple types of records in the same name entry was also tested and verified.

Change-Id: I4cbac5bad85daf3a0d2c0cc403372e34f5150c67