Update VMDNS Init Process and Cache Refresh Thread
Before, the VMDNS init process set the service state to READY when the
cache was able to load initial zone data from the directory.  The point
where the state was set caused some timing conflicts with other services
that relied upon VMDNS to be ready.  Also, the VMDNS cache refresh
thread was not that reslient - it would exit if a function call resulted
in an error.  When the cache refresh thread died, the cache would cease
to synchronize causing it to have improper data unless a manual refresh
was done.

This change updates the VMDNS init process such that the initialization
function waits until the VMDNS state is set to READY before finishing.
It will repeatedly, there is a timeout of five attempts,  signal the
cache refresh thread so it can attempt to propogate initial data and
set the VMDNS state to ready.

This change also prevents the cache refresh thread from bailing when a
function call fails.  Instead, all errors are logged and the cache
refresh thread tries again to synchronize data.  The only instance that
the cache refresh thread can fail is when the timed wait fails.

Changes made:

  - Replaced all, but one, BAILs from VmDnsCacheRefreshThread with error logging
statements
  - Added a loop and a condition signal to VmDnsSrvInitDomain to run
VmDnsCacheRefreshThread to propogate initial data from the directory and
set the VMDNS state to READY
  - Removed hacky solution to circumvent this issue in VMAFD DNS Init
sequence

Test plan:

  - With changes from
(781)[http://code.pslabs.eng.vmware.com:8082/#/c/781/], VMAFD was not
able to initialize reverse DNS zones due to a timing mismatch bewteen
VMAFD's thread and the VMDNS cache thread setting the service state to
READY
    + Adding this change mitigates the failure and VMAFD is able to
initialize reverse DNS zones
  - Before, cache refresh thread would exit if the server state was not
ready and its function calls would fail
    + Adding this change does not cause the cache refresh thread to
exit, instead, it logs errors and retries

Change-Id: I4e949685264133d5329611c894aab8c4e76d6a40