Refactor RPM packages
This change refactors the RPM packages as follows:
  lightwave         (contains vmware-sts)
  lightwave-server  (contains vmdir, vmdns, vmca)
  lightwave-client  (contains client libraries and vmafd)
  lightwave-post    (contains post nee lwraft)
  lightwave-devel   (contains includes files)

I tested this on photon with the following package installs:
  lightwave-client, lightwave-server: ic-promote works
  lightwave-client, lightwave-server, lightwave: configure-lightwave-server works

To get configure-lightwave-server to succeed I fixed several issues
in vmidentity. It was checking for the package name vmware-sts; it now
checks for the name lightwave. Also the jsvc command was not set correctly,
and the systemd file vmware-stsd.service needed a change to the exit
success value.

Since the automated build will be changed to use these new packages, and
the old build method will be retired, I have consolidated the configure.ac
files for each subdirectory into a single top-level configure.ac and modified
the Makefile.am files to reference the appropriate .la files. This makes
the top level build much more usable and consistent with conventional autotools
practice.

PR#1910715, PR#1913710

Change-Id: I0fee39b3435b6f8e535dd69945fa7316f83dc83c