Change VMDNS Cache Thread Logging
Before, the VMDNS cache refresh thread would always print out, and
immediately try again it could not refresh.  This behavior would flush
syslog and consume too much CPU.

Now, the cache refresh thread has been changed so it sleep and waits for
five seconds upon failure, and also changes the log level of the
statement that flooded syslog such that it is only present if the user
desires.

This solves bug
[1926963](https://bugzilla.eng.vmware.com/show_bug.cgi?id=1926963)

Change plan:

  - Coditional wait if the VMDNS CacheRefreshThread cannot communicate
with/query the directory
  - Change log level of output if the CacheRefreshThread cannot
communicate with/query the directory

Test plan:

  - Ran Lightwave and did not promote
  - Before, syslog would become flooded
  - Now, that does _not_ happen and VMDNS does not consume max CPU

Change-Id: Ic93d747293cdf446677276d3b913d395bbba2c20