Collect VMDIR LDAP metrics using metrics API
With this update, vmdir ldap request durations for different operations and
ldap error counts are collected and maintained using the metrics API.

There are some changes to vmdir rest-head to get these metrics in prometheus format on HTTP GET request.

The update also includes some changes to Metrics API to get the data in expected format.

Change-Id: I5cab22ca9a751b151eaeee0b9a5299f3f71eabee