start vamfd service during ic-join
When ic-join was run without the --domain-controllers argument,
it tried make a call to the vmafd service to validate domain
credentials. This call was made prior to starting the vamfd
service, so it failed.  Move the validation of domain credentials
following the startup of vmafd service.

Also updated the RPM postinstall script to refresh lwsm after
importing registry keys.

PR#1928213

Change-Id: I3a496f0c0c1e2074b639ef733e904341d6cb14b5