Fixed bug causing vmdnsd to crash on shutdown and added api endpoints for deregistering from and freeing event queue.
Before, vmdns crashed when shutting down due to a race condition where
shutdown thread freed memory other threads used before exiting. Also,
there was a method to add a fd to event queue, but none to remove one.
CloseEventQueue() did not free the queue as it claimed. It is renamed to
ShutdownEventQueue and there is a separate method to free queue.

Fixed use after free race condition. The shutdown thread is now
responsible for closing and releasing the file descriptors epoll_wait
is waiting on. On close, worker threads will check bShutdown flag and
exit. Refactored code so that all fd's are removed from queue before
closing and freeing.

Changes:
- Created method for freeing EventQueue and added it to VmSock api
because CloseEventQueue() did not free it.
- Created method for removing from EventQueue, similar to AddEventQueue
- Renamed CloseEventQueue() to ShutdownEventQueue()
- On EPOLLHUP or when closing a thread, the sock is detached from event
queue
- Added flag in PVM_SOCKET to check if it is in an event queue
- Thread initiating shutdown will shutdown queue, remove sockets from
the queue, and then close them. At the end, the queue is freed.

Testing:
- ran with 'lwsm gdb vmdnsd' and ensured the process exitted normally when
triggering a shutdown with 'lwsm stop vmdns' and sending SIGKILL.
- Also ran 'lwsm shutdown' and ensured 'lwsm start vmdns' worked.
- Ran with more than default (4) threads and tested shutdown
- tested with dig +tcp to see if sockets were being opened properly

Change-Id: I6705e3456b06644111b1754677b4231c5647b326