PR 1955016: Avoid unnecessary replication cycle retry, if lowestPendingUncommittedUsn case is hit
Problem:
If the USN, consumer is trying to pull is lowestPendingUncommittedUsn then consumer ends up sending high watermark in syncdone control which results in replication retry.

Changes made:
In this case replication retry is not necessary, modified the code to ensure we return lowestpendingcommittedUSN as part of sync done control in the case of continue.

Testing performed:
Tested on a three node ring topology and verified that consumer pulls the changes correctly if lowestpendingcommittedUSN case is hit.

Change-Id: I3af0e153b57edaef71ebd92823bc591cbcace801