Added CI/CD deployment pipeline scripts
The appspec.yml file will list which scripts will run on each AWS
defined event. The scripts will upgrade lightwave or install it in a
fresh instance and configure it appropriately (as a DC or replication
partner). If there already is a DC, then it will be configured as a
partner.

These files will be uploaded to a pre-defined S3 bucket (along with the
built rpms) to trigger the pipeline.

Testing:
- Set up auto-scaling group to deploy one node and run script. Verified
that the instance was sucessfully promoted to DC.
- Modified auto-scaling group to deploy 2 nodes max. Verified that the
newly deployed instance was configured as a replication partner to the
first instance (Domain Controller).
- Tested pipeline by Releasing change and verifying Alpha-Test was
successful
- Uploaded S3 bucket with newer rpms and verified that upgrade was
triggered.

Change-Id: I56ad3102c8f95d3d1d84cb8d47dc61a6b256b96d