Promoting Partner Creates All DNS Records
Bug Number: 1968860

There was a regression introduced which did not create a partner node's
SRV and NS records upon promotion.  To circumvent this, a user would
have to run configure-lightwave-server twice, where the second time, the
`-n` flag would have to be used to create the proper DNS entries for the
partner node.

With this fix, promoting the partner node results in having all of its
DNS entries created.

Change list:

  - Update the VmAfSrvConfigureDNSW and VmAfSrvConfigureDNSA function
signatures to take an additional argument - TargetName.  TargetName
configures the entries on the specified target.
  - Update all uses of these functions
  - Update the promotion sequence to get the system hostname through the
PNID value stores in the regitry.  Before, partner node promotion
obtained the hostname through the system `gethostname` utility.  But
there is a case where you can promote with a different name and this
value is stored in the registry's PNID field.
  - Add dns-entries-test.sh script to verify DNS record creation for
federations

Testing:

  - Promote standalone node and ensure DNS entries (1 SOA, 4 SRV, 1 A,
1 NS) are correctly created
  - Promote partner node and ensure all DNS entries (4 SRV, 1 A, 1 NS)
are created
  - Join a client node and ensure all DNS entries are created (1 A)
  - Leave the client node and ensure its A record is deleted
  - Demote the partner and ensure its 6 DNS entries are deleted
  - Use the test script in `vmafd/test/scripts/dns-entries-test.sh` to
verify results

Change-Id: Icd1d9a6f480b3fc00d78a8b24a5981fe9116421e