POST Logging Same workflow step executed twice
This change is to add more logging to the post rest modify request.
Initially the request logged only the dn being modified, with this
change the content of the mod request will be logged too.

Bug Number: 1976965

Tests done:
Created a local post instance and issues a modify (PATCH) request.
Confirmed that the logs show the contents. Maximum 256 value per
attribute are logged and length of each value is capped to 256.

Change-Id: Ic4658ae06208fd94b35321e74a59a48c9f4350bb