Do Not Forward DNS Request If No Forwarders
Before, if there are no forwarders configured, VMDNS would still try to
call the forward request function which would result in an error.

This error would make response times slow, and it would also result in
an improper response packet.  Fuzz tests failed if there were no
forwardrs configured and we tried to test VMDNS.

To fix this, we call the ForwardRequest function if there is an error
generated from attempting to fulfill the request locally from VMDNS
zones.  If ForwardRequest cannot fulfill the request (for example, if
there are no forwarders configured), then it will immediately notify the
sockinterface function processing the TCP/UDP request.  If there is an
error from processing the request locally, or if the ForwardRequest
function generated an error, we will send the response from our local
request handling.

Changes:

  - sockinterface.c calls ForwardRequest if there is an error generated
from local processing, if ForwardRequest generates an error or if local
processing generates an error, then we send the response generated from
local processing.

Testing:

  - Run VMDNS with no forwarders configured
  - Run the fuzzer which went through this issue
  - Response was correctly returned if manually attempting to resolve
non-existant entry

Change-Id: Icdfa2e66812e224b3edf5c4cea643979806ea3e1