Lets VMDNS Send Proper Error Response
There was a regression introduced in VMDNS with c04d4761 where if local
processing of a request failed, then we would immediately bail out of
the sockinterface function for UDP/TCP.  This would prevent VMDNS from
sending proper error responses if it cannot resolve the request.

Also, there were some logical errors in processing error codes to send
the appropriate response from the sockinterface layer.

To address this, we remove the bail line after ProcessRequest and allow
the sockinterface functions to try forwarding or sending error
responses.  Also, we fix the logic when processing error codes from our
local ProcessRequest function and the Forwarder function.

Testing:

  - VMDNS with no forwarders configured
    + Try to resolve a known domain name and expect an answer saying
error
      - Got an answer saying error
    + Try to resolve a domain name in the default zone and expect an
answer with success
      - Got an answer with success

Change-Id: Id9dbc48465013a8d05ee51daac8664b7f07ee0c3