VMDNS Does Not Forward If Query Within Zone
Before, if a query failed to resolve inside our zones, VMDNS would
attempt to forward it.  This is incorrect behavior.  VMDNS should act
authoritatively for queries within one of its zones.

This lead to unncessary forwarding requets which opened unncessary FDs
on the host.  For example, if a query for a AAAA record in our zone does
not exist, then previously, VMDNS would try to forward it.  The
forwarder would obviously not be able to resolve it.

This change adds an additional check to the sockinterface layer which
does not forward the request if the ProtocolHandler says that it was a
query within one of our zones.

Changes:

  - VmDnsOnTcpRequestDataRead and VmDnsOnUdpRequestRead both check to
see if the protocol handler's entry function, VmDnsProcessRequest,
returns that the request was a query within one of VMDNS's zones
  - VmDnsProcessRequest returns an additional value - bQueryInZone which
tells the sockinterface layer that this request was a query in one of
VMDNS's zones
  - VmDnsProcessQuery returns an additional value - bQueryInZone which
tells VmDnsProcessRequest that this request was a query in one of
VMDNS's zones

Testing:

  - Sniffed the network on the Lightwave node on port 53 while trying to
resolve a AAAA record for the hostname of the server
    + There is no AAAA record for the hostname in the zone
    + Expected behavior is that NXDomain is returned and the network
makes no requests to the configured forwarder to resolve the request
    + Actual behavior lined up with the expected behavior - response was
NXDomain and there was no traffic to the forwarder to resolve this
request

Change-Id: I583d3f5ee42a96b60a0e41ec8ed075ee4c015368