vmdird/vmdnsd/vmcad logging to syslog with wrong names
Bug: 1997061

Problem:
vmdird/vmdnsd/vmcd/post services are now logging to syslog with the wrong names.
For example, vmdird messages are now logged with the name "vmdird-https".
These should be changed back to the appropriate names, "vmdird", "vmdnsd" etc.
In addition to that journald rate limit is high, LOG only in the case of error.

Change:
Changed all the daemon name provided to the c-rest-engine config.
Since latest version of c-rest-engine (1.1-1) makes use of syslog,
debug log file name is not valid. Removed the debug log file provided
to c-rest-engine config.

Test:
Verified that with this change, all daemon name are displayed corrrectly in
syslog.

Change-Id: I214c5fb22b8794586de0083b1ceceb1ad8809fb7