Bypass HOK Token POP Validation VMDIR REST Head
Currently, the VMDIR REST head does not validate the signed HTTP request
that comes with HOK tokens.  Instead of accepting the request, the REST
head just bails if the token type is HOK.

The token validation code correctly verifies HOK tokens, but there does
not exist code to validate the signature.  To unblock services that need
to communicate with the VMDIR REST head using HOK tokens, this change
temporarily bypasses signature validation and only validates the HOK
token itself.  PR 2004701 can track the status of adding signature
validation to the VMDIR REST head.

Once the signature validation method is added, there should be a
transparent change to end users of the VMDIR REST head.

Testing:

  - Ran a cURL request against various VMDIR REST APIs using HOK token
as the auth method.

Change-Id: Iea866aad12303e70545fe5b5c173278e45571587