Set DCName to FQDN when promoting directory
This change sets the DCName registry parameter to be
the FQDN of the host when promoting a directory server.
It was being set "localhost" which caused kerberos
authentication from the vmdir auth provider to fail.

Also added a test tool lwraft_test_access that
creates an access token for a given user.

PR#1926648

Change-Id: I3d2d68fa93ed1bc4cdf86131b3fd2ed76529462d