Implements Region aware ic-join and promotion
Before, ic-promote did not add sitename to the SRV records on promotion.
Also, domainjoin did not specify a region to join to.

Now, ic-promote will add SRV records specifying region if the --site
option is enabled. ic-join will have a --site argument to join to a DC
at a specific site. If the site is not found or no DC exists in the
site, ic-join will fail. Running ic-join with no site argument will join
it to a DC in the specified domain, regardless of site. The server will
remain backwards compatible with older clients by supporting the old
ic-promote and demote api (where site is not passed in)

Changes:
- Adds new api such that on promotion, sitename (optional argument) is
 also sent via rpc to vmdns. Same is done for demote api.
- If a sitename is included on promotion, an SRV record will be created
for the specific site/region, as well as another SRV record for the
domain in general
- Modified ic-join cli to include --site option, searches SRV records
for specific site
- Fixed bug on ic-join where failing to join to a DC reported success
and did not bail on error
- Cleans up site specific SRV records on demote

Testing Done:
- ic-promote without site on one LW instance. On another, try ic-join
with and without region specified. ic-join with region will fail,
without region will succeed.
- ic-promote with site on one LW instance. On another instance, promote
as replication partner for another site on same domain. On another
instance, try ic-join with and without region specified, verify both calls
succeed. Verify ic-join with incorrect site fails.
- list vmdns records between all cases and verify promote and ic-join
adds correct records and domainjoin leave removes A record
- domainjoin for all ic-join test cases
- demote replication partner, verify SRV records are properly purged

Change-Id: Ibbd0dca995adbafffa4f43a1f76ce9965ab5ebc0