ZERO Downtime POST - Part 2 (revision 2)
This is the second part of the change for zero downtime post
This change was made to add the retry logic for REST requests.
With this code we retry a request if the error code returned is
a retriable error code (503/408/599). We maintain the request till
either the retries reach max number or the error code is not retriable.

Change-Id: Idd3c08fbdcfaf08b00c909a8880c6ef80334ffd2