Code deploy changes for zero downtime POST/ Latest c-rest-engine
This chnage was made to add the code deploy changes for zero downtime.
With this change we now only upgrade a follower node. If a node is a
leader we trigger the startvote tool to elect a new leader and then
update the current node.

Also made changes to update to latest c-rest-engine

Change-Id: I695b20293a7112744058d44e348b6414bc56cedb