Add connection retry logic to client join
If ic-join fails to connect to vmdir while validating credentials,
retry the connection up to 2 more times with a delay of 3 seconds
between retries. Also updated the retry logic for domain join to
retry for only specific error codes and to delay for 1 second before
retrying.

Testing: Ran ic-join to a DC where vmdir was disabled. It retried
the connection twice before reporting failure. Ran it again
and started vmdir after the first connection attempt--the join
succeeded on the second retry.

Update RPM patch level to 10.

PR#2010780

Change-Id: Iddc20bd8eb023e5f99cffccd0ee0163d1bec3978