Adds go wrapper to getDCName with force refresh flag
As part of Lightwave HA, a go wrapper will be needed to get the DC name
from vmafd and force a refresh. This change adds that api so that the
returned DC will be live with regards to the heartbeat check.

The test for ForceRefresh will not succeed currently because
ForceRefresh does not force re-affinitization of a DC. If a service is
down and ForceRefresh is called, the state machine will wake up and
query vmafd server, which will report the node is still live since a
couple more failed heartbeats are still needed. This will be addressed
in a future CR.

Testing:
- Manually verified with gdb that getDCName was called with force
refresh flag and state machine woke up and checked heartbeat

Change-Id: Ib5e350d0f4c71c804ec9969037a09c386d5ff324