Adds Safety Before Dereferencing From DNS Buffer
Previously, parsing strings from a DNS buffer did not check the buffer
length when attempting to grab the label length from the buffer.
Because of this, there is a case where a malformed packet could be
processed correctly and an impropper length could be used to further
parse the DNS packet.

This change adds a check to ensure that you can pull a byte out of the
DNS buffer containing the label length.  If you can do that, then the
code will proceed to parse the rest of the string from the DNS buffer.

This change remediates a possible heap buffer overflow in VMDNS.

Bug Number: 2017926

Changes:

  - Update buffer.c function to read string at an offset to validate the
buffer size before dereferencing data from it

Testing:

  - Ran fuzzer and performance test against VMDNS (compiled with address
sanitizer) without my change and got a buffer overflow segfault
  - Ran fuzzer and performance test against VMDNS with my change and got
no buffer overflow
  - Ran additional fuzzer with exhaustive DNS checks

Change-Id: Ie267bd6bdfcdfbb37b75db1a9f64e91e9e539521