Fixes SRV record query in vmafd cdc layer to use FQDN
In the CDC client layer, an SRV record query was sent to vmdns with a
domain name not including the dot '.' at the end. Since zones are stored
with the trailing dot, no DCs were found. This change adds the dot at
the end of the domain so the correct zone is queried.

Changes:
- added '.' at end of domain so correct FQDN was queried when finding DCs
- added test script for verification

Testing:
- tested with client against 2 node federation
- ran cdc-cli cache list and verified correct DCs were found
- ran test script

Change-Id: I7b30fd6bb52e87e33ad3a913f465d4f9988b19af